split expandable and non-expandable section links into two components
